useTextEditor 커스텀훅 제작
- TextEditor의 기능을 커스텀훅으로 분리
